+++ src/components/add_to_cart_btn/add_to_cart_btn.js
@@ -4,27 +4,24 @@ import { faCartPlus } from '@fortawesome/free-solid-svg-icons';
 import { Button } from 'reactstrap';
 import { CartManager } from '../catalog_page.js';
 
-export default class ToCartBtn extends React.Component {
+export default class AddToCartBtn extends React.Component {
   render() {
     const { product, quantity, dropQuantity } = this.props
     return(
       <CartManager.Consumer>
-        {
-          manager => {
-            return (
-              <Button
-                color='success'
-                onClick={(e) => {
-                    manager.addToCart(product, quantity)
-                    dropQuantity()
-                  }
+        { 
+          manager => 
+            <Button
+              color='success'
+              onClick={(e) => {
+                  manager.addToCart(product, quantity)
+                  dropQuantity()
                 }
-              >
+              }
+            >
               <span>В корзину</span>
               <FontAwesomeIcon icon={ faCartPlus }  size="1x"/>
             </Button>
-            )
-          }
         }
       </CartManager.Consumer>
     )
